lms/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-3-jet==1.0.8
django-froala-editor==3.2.2
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
